+++ main.py
@@ -7,58 +7,16 @@
 # Input file one line: Name, Surname, balance, loan
 # • Write transactions in to a file called “Transaction.txt”
 import os
+from app.validation_check import get_number
+from app.file_usage import get_data_from_file
+
+# my_directory = os.path.dirname(os.getcwd())
+# print(my_directory)
 
 ################################################################### store expenses and balances in empty lists
 expenses_sheet = []
 balances_list = []
-#################################################################### get valid number from user
-def get_number(user_input, error_ms):
-    while True:
-        user_input_number = input(user_input).strip()
-        if user_input_number.replace(".", "", 1).isnumeric():
-            user_input_number = float(user_input_number) if '.' in user_input_number else int(user_input_number)
-            return user_input_number
-        else:
-            print(error_ms)
-############################################################################# store data from user in function file_path
-def get_data_from_file(file_path):
-    try:
-        with open(file_path, "r") as file_data:
-            ################################################################### read user's name and surname form the provided file
-            name = file_data.readline().strip()
-            if len(name) == 0:
-                name = input("Name is missing in the provided file. \n Enter user's name manually: ")
-            surname = file_data.readline().strip()
-            if len(surname) == 0:
-                surname = input("Surname is missing in the provided file. \n Enter user's surname manually: ")
-            ################### receive current balance from the provided file, if data is incorrect, ask to provide it manually
-            balance_file = file_data.readline().strip()
-            if balance_file.replace(".", "", 1).isnumeric():
-                balance = float(balance_file) if '.' in balance_file else int(balance_file)
-                print("The Current Balance from the provided file is:", balance)
-            else:
-                print("The Balance is invalid or missing in the provided file.")
-                balance = get_number(user_input="Enter user's current balance manually: ",
-                                     error_ms="Please enter numeric, positive value for balance: ")
-            ################### receive loan balance from the provided file, if data is incorrect, ask to provide it manually
-            loan_file = file_data.readline().strip()
-            if loan_file.replace(".", "", 1).isnumeric():
-                loan_balance = float(loan_file) if '.' in loan_file else int(loan_file)
-                print("The current Loan Balance from the provided file is:", loan_balance)
-            else:
-                print("The Loan Balance is invalid or missing in the provided file.")
-                loan_balance = get_number(user_input="Enter user's loan balance manually: ",
-                                          error_ms="Please enter numeric, positive value for loan balance: ")
-                ################# ask user to input their first expense
-            expense = get_number(user_input="Enter user's first expense manually: ",
-                                     error_ms="Please enter numeric, positive value for expense: ")
-            return name, surname, balance, loan_balance, expense
-    except FileNotFoundError as error:
-        print(f"Provided file doesn't exist", {error})
-        return None
-    except Exception as error:
-        print(f"An error occurred:", {error})
-        return None
+
 #################################################################### ask for data input choice
 while True:
     select_input = input("Do you want to enter data manually? yes/no: ").lower().strip()
@@ -82,19 +40,21 @@ else:
     while True:
         # file_path = r"C:\projects\Budget_Tracker\CustomerData.txt"  ##### for testing use
         path_input = input("Enter file's path: ").strip()
+        ############################################################### validate the inputted file's path
         file_name = "CustomerData.txt"
         if os.path.basename(path_input) == file_name:
             file_path = path_input
         else:
             file_path = os.path.join(path_input, file_name)
+        ############################################################## get data from the file
         file_data = get_data_from_file(file_path)
         if file_data:
             name, surname, balance, loan_balance, expense = file_data
             break
         else:
             exit()
-################################################################### remove extra spaces and get full name
 
+################################################################### remove extra spaces and get full name
 full_name = name.title() + " " + surname.title()
 full_name_split = full_name.split()
 full_name = " ".join(full_name_split)
@@ -135,6 +95,7 @@ while not done:
     balances_tuple = tuple((balance, loan_balance))
     ##################################################################### convert balance data to a list, so it can be added more
     balances_list.append(balances_tuple)
+
     ##################################################################### write transaction data into separate file
     with open("Transaction.txt", "w") as transaction_file:
         for i in range(len(expenses_sheet)):
@@ -146,6 +107,7 @@ while not done:
                                    f"Expense: {expense_in_file}, Balance: {balance_in_file}, Loan balance: {loan_in_file},"
                                    f"Expense ID: {expense_ID_in_file}.\n")
     print("The transaction data is written to 'Transactions.txt'.")
+
     ##################################################################### ask for a new expense
     while True:
         next_expense_prompt = input("Would you like to do another expense? (yes/no) ").lower().strip()
@@ -169,3 +131,5 @@ else:
         print(f"{i+1}) Expense: {expenses_sheet[i]["expense_count"]},"
               f" Balance: {balances_list[i][0]}, Loan: {balances_list[i][1]}."
               f"\n  Expense ID is: {expenses_sheet[i]['expense_id']}.")
+
+
